move code to datasource package
